# cmd/filmlibrary/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Film Library Application",
        "title": "Film Library API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/films/{id}": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Delete a film by path parameter 'id'",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "Delete a film by path parameter 'id'",
                "operationId": "delete-film",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Film`s id that needs to be deleted",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_entrypoint_http.deleteFilmResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Update fully or partially a film by path parameter 'id'",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "Update fully or partially a film by path parameter 'id'",
                "operationId": "update-film",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Films`s id that needs to be updated",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Film object with values that will be updated",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_entrypoint_http.updateFilmRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_entrypoint_http.updateFilmResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "internal_app_entrypoint_http.deleteFilmResponseBody": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_app_entrypoint_http.filmActor": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sex": {
                    "type": "integer"
                }
            }
        },
...
        "internal_app_entrypoint_http.updateFilmRequestBody": {
            "type": "object",
            "properties": {
                "actor_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1,
                        2,
                        3
                    ]
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "rating": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 0
                },
                "release_date": {
                    "type": "string",
                    "example": "2006-01-02"
                },
                "title": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                }
            }
        },
        "internal_app_entrypoint_http.updateFilmResponseBody": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_app_entrypoint_http.filmActor"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}